fix: jest config
