--- .jest.setup.js
@@ -1,20 +1,0 @@
-// Optional: configure or set up a testing framework before each test.
-// If you delete this file, remove `setupFilesAfterEnv` from `jest.config.js`
-
-// Used for __tests__/testing-library.js
-// Learn more: https://github.com/testing-library/jest-dom
-import '@testing-library/jest-dom/extend-expect';
-
-const nextJest = require('next/jest');
-
-const createJestConfig = nextJest({
-    dir: './',
-});
-
-const customJestConfig = {
-    setupFilesAfterEnv: ['<rootDir>/jest.setup.js'],
-    moduleDirectories: ['node_modules', '<rootDir>/'],
-    testEnvironment: 'jest-environment-jsdom',
-};
-
-module.exports = createJestConfig(customJestConfig);
