+++ package.json
@@ -35,17 +35,21 @@
   "devDependencies": {
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^14.4.3",
     "@types/node": "18.11.18",
     "@types/react": "18.0.27",
+    "@types/testing-library__jest-dom": "^5.14.5",
     "@typescript-eslint/eslint-plugin": "^5.49.0",
     "autoprefixer": "^10.4.13",
     "eslint": "^8.32.0",
     "eslint-config-next": "13.1.5",
+    "eslint-plugin-jest-dom": "^4.0.3",
     "jest": "^29.4.1",
     "jest-environment-jsdom": "^29.4.1",
     "postcss": "^8.4.21",
     "sass": "^1.57.1",
     "tailwindcss": "^3.2.4",
+    "ts-jest": "^29.0.5",
     "typescript": "^4.9.4"
   }
 }
